Move entities only used by V2 resources to entities/v2.
Entities moved: `ConsumerAssignmentRequest`, `ConsumerAssignmentResponse`, `ConsumerCommittedRequest`, `ConsumerCommittedResponse`, `ConsumerOffsetCommitRequest`, `ConsumerSeekToOffsetRequest`, `ConsumerSeekToRequest`, `ConsumerSubscriptionRecord`, `ConsumerSubscriptionResponse`, `TopicPartition`, `TopicPartitionOffsetMetadata` and `TopicPartitionOffsetResponse`.
