+++ kafka-rest/src/main/java/io/confluent/kafkarest/entities/v2/ConsumerAssignmentResponse.java
@@ -13,14 +13,14 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.kafkarest.entities;
+package io.confluent.kafkarest.entities.v2;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.List;
 
-public class ConsumerSubscriptionResponse {
+public class ConsumerAssignmentResponse {
 
   @JsonProperty
-  public List<String> topics;
+  public List<TopicPartition> partitions;
+
 }
