+++ kafka-rest/src/main/java/io/confluent/kafkarest/entities/v2/ConsumerOffsetCommitRequest.java
@@ -13,13 +13,19 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.kafkarest.entities;
+package io.confluent.kafkarest.entities.v2;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.List;
 
-public class ConsumerCommittedResponse {
+public class ConsumerOffsetCommitRequest {
+
+  @JsonCreator
+  public ConsumerOffsetCommitRequest(@JsonProperty("offsets") List<TopicPartitionOffsetMetadata>
+      offsets) {
+    this.offsets = offsets;
+  }
 
   @JsonProperty
   public List<TopicPartitionOffsetMetadata> offsets;
