+++ kafka-rest/src/main/java/io/confluent/kafkarest/entities/v2/ConsumerSeekToRequest.java
@@ -13,13 +13,12 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.kafkarest.entities;
+package io.confluent.kafkarest.entities.v2;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.List;
 
-public class ConsumerAssignmentResponse {
+public class ConsumerSeekToRequest {
 
   @JsonProperty
   public List<TopicPartition> partitions;
