+++ kafka-rest/src/main/java/io/confluent/kafkarest/entities/v2/TopicPartition.java
@@ -13,34 +13,24 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.kafkarest.entities;
+package io.confluent.kafkarest.entities.v2;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
+import javax.validation.constraints.Min;
 import org.hibernate.validator.constraints.NotEmpty;
 
-import java.util.Objects;
-import javax.validation.constraints.Min;
-
-public class TopicPartitionOffsetMetadata {
+public class TopicPartition {
 
   @NotEmpty
   private String topic;
   @Min(0)
   private int partition;
-  @Min(0)
-  private long offset;
-
-  private String metadata;
-
-  public TopicPartitionOffsetMetadata(
-      @JsonProperty("topic") String topic, @JsonProperty("partition") int partition,
-      @JsonProperty("offset") long offset, @JsonProperty("metadata") String metadata
+    
+  public TopicPartition(
+      @JsonProperty("topic") String topic, @JsonProperty("partition") int partition
   ) {
     this.topic = topic;
     this.partition = partition;
-    this.offset = offset;
-    this.metadata = metadata;
   }
 
   @JsonProperty
@@ -63,26 +53,6 @@ public class TopicPartitionOffsetMetadata {
     this.partition = partition;
   }
 
-  @JsonProperty
-  public long getOffset() {
-    return offset;
-  }
-
-  @JsonProperty
-  public void setOffset(long offset) {
-    this.offset = offset;
-  }
-
-  @JsonProperty
-  public String getMetadata() {
-    return metadata;
-  }
-
-  @JsonProperty
-  public void setMetadata(String metadata) {
-    this.metadata = metadata;
-  }
-
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -92,14 +62,8 @@ public class TopicPartitionOffsetMetadata {
       return false;
     }
 
-    TopicPartitionOffsetMetadata that = (TopicPartitionOffsetMetadata) o;
+    TopicPartition that = (TopicPartition) o;
 
-    if (offset != that.offset) {
-      return false;
-    }
-    if (!Objects.equals(metadata, that.metadata)) {
-      return false;
-    }
     if (partition != that.partition) {
       return false;
     }
@@ -114,7 +78,6 @@ public class TopicPartitionOffsetMetadata {
   public int hashCode() {
     int result = topic != null ? topic.hashCode() : 0;
     result = 31 * result + partition;
-    result = 31 * result + (int) (offset ^ (offset >>> 32));
     return result;
   }
 }
